Extract `OsType` into own file
